debug uploading and removing avatars
